refactor: ensure float type consistency in DLMOCalculator and MoodPredictor
This commit updates the DLMOCalculator to append hours as floats in the cbt_rhythm list, enhancing type consistency. Additionally, the MoodPredictor is modified to ensure that model confidence values and model information are returned as floats and properly typed, respectively. These changes contribute to the overall goal of a high-performance, human-centered psychiatry platform through clean and scalable code practices.

+++ src/big_mood_detector/domain/services/dlmo_calculator.py
@@ -327,7 +327,7 @@ class DLMOCalculator:
             # CBT is proportional to circadian state
             # Note: This simple proxy results in ~6h phase delay
             cbt = -state_xc * math.cos(state_x)
-            cbt_rhythm.append((hour, cbt))
+            cbt_rhythm.append((float(hour), cbt))
 
         return cbt_rhythm
 
@@ -417,7 +417,7 @@ class DLMOCalculator:
         Converts light intensity to circadian drive.
         """
         # Hill equation: alpha = a0 * (I^p / (I^p + I0^p))
-        return self.A0 * (light**self.P / (light**self.P + self.I0**self.P))
+        return float(self.A0 * (light**self.P / (light**self.P + self.I0**self.P)))
 
     def _extract_dlmo_enhanced(
         self, cbt_rhythm: list[tuple[float, float]], target_date: date
